requirements_onnx.txt
# Ultra-minimal requirements using ONNX instead of PyTorch
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Image processing - minimal versions
opencv-python-headless==4.8.1.78
numpy==1.24.4
pillow==10.0.0

# ONNX runtime (much smaller than PyTorch)
onnxruntime==1.16.3

# Database and utilities
supabase==2.7.0
python-dotenv==1.0.0
